+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+## [0.16.0]
+         
+- Added: introduce zt-exec to call JBang command
+- Added: introduce ProgressManager and Task.Backgroundable to sync dependencies asynchronously
 - Fix: added descriptions to directive completions
 
 ## [0.15.0]